refactor(service): session files in separate module
