+++ data-plane/core/service/src/session/config.rs
@@ -1,0 +1,20 @@
+// Copyright AGNTCY Contributors (https://github.com/agntcy)
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::session::multicast::MulticastConfiguration;
+use crate::session::point_to_point::PointToPointConfiguration;
+
+#[derive(Clone, PartialEq, Debug)]
+pub enum SessionConfig {
+    PointToPoint(PointToPointConfiguration),
+    Multicast(MulticastConfiguration),
+}
+
+impl std::fmt::Display for SessionConfig {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            SessionConfig::PointToPoint(ff) => write!(f, "{}", ff),
+            SessionConfig::Multicast(s) => write!(f, "{}", s),
+        }
+    }
+}
