+++ data-plane/core/service/src/session/moderator_task.rs
@@ -1,9 +1,11 @@
 // Copyright AGNTCY Contributors (https://github.com/agntcy)
 // SPDX-License-Identifier: Apache-2.0
 
+// Third-party crates
 use tracing::debug;
 
-use crate::errors::SessionError;
+// Local crate
+use crate::session::errors::SessionError;
 
 #[derive(Debug, Default)]
 pub(crate) struct State {
